+++ packages/studio-core/src/index.ts
@@ -58,6 +58,7 @@ export * from './lib/chordService';
 export * from './lib/updater/updaterSimulation';
 export * from './lib/updater/diagnostics';
 export { deleteLocalApk } from './lib/updater/cacheManager';
+export * from './lib/updater/versionLogger';
 export * from './lib/updater/stateMachine';
 export * from './lib/startupCoordinator';
 export * from './lib/themeEngine';